refactor(storage): keep lane forwarder internal
